// frontend/src/types/weather.ts
export interface AQIInfo {
  aqi: number;
  status: string;
  color: string;
  source: string;
  pm25?: number;
  pm10?: number;
  no2?: number;
  so2?: number;
  co?: number;
  o3?: number;
}

export interface CurrentWeather {
  location: string;
  district: string;
  state: string;
  date_str: string;
  updated_at: string;
  temperature: number;
  feels_like: number;
  maximum: number;
  minimum: number;
  humidity: number;
  wind_speed: number;
  wind_direction: string;
  wind_direction_deg: number;
  condition: string;
  icon: string;
  aqi: AQIInfo;
}
// ...
export interface FavouriteItem {
  id: number;
  location_name: string;
  district: string;
  state: string;
  latitude: number;
  longitude: number;
  current_temp: number;
  min_temp: number;
  max_temp: number;
  condition: string;
  created_at?: string;
}